Log improvements and bugfixes
* Add logging facilities

All log messages will be prefixed with a timestamp, a log type and
will be printed to stdout.

Current available types are: DEBUG, INFO and ERROR.
DEBUG messages can be disabled at runtime by setting DEBUG_ON to false.

* fix exception in validSimpleCapture

If fs.open failed, both the error path and the success path where taken
because of the missing else, causing an uncaught exception.

* bump version number to 0.1.1

* enforce launchPlayer is executed as root

* force launchPlayer's exit code match driver's exit code

This change allows us to check the exit code of the driver.
Thr script always returned 0 because a bash script's exit code
is the same as the last executed command. In this case, it was *sleep*
which it always returned 0.

* reset FPGA at end of launchPlayer

Buy issuing always the reset, even if the reproduction ended with
an error we know the FPGA will always be in the IDLE state.

* bump version number to 0.1.2

* add to all exported functions displayName attribute

Added a function in _common 'prettyName' to generate the
displayName using the path to the file and a string which should
be the function's name.

Functions that call logDebug, logError or log will be expected to have
the attribute prettyName.

* swap console.log in server.js to common.log

Start using new logging utilities.
Modified the main logging functions. Know the messages are prefixed
with the caller's prettyName.

* fix comments in log and logError

My initial intention was to rewrite console.log and console.error
but I left the old comments.

* change debug messages to common.logDebug2

Added new function logDebug2 to use in anonymous callbacks and
functions that don't have defined the prettyName attribute.

* add some debug messages

* bump version to 0.1.3

* fix ReferenceError exception

* bump version number to 0.1.4

* fix card2host and host2card stopping commands

To correctly stop those programs, one has to send SIGINT, SIGTSTP and
then SIGCONT.

Thanks @jfzazo for the tip.

* bump version number to 0.1.5

* FIX: MISSING RETURN CAUSED WEIRD APP CRASH

* Bump version number to 0.1.6

* fix: launchPlayer status is always 0

status held the return vale of sleep, which was always 0, and that's
not the return value I intended to catch.

* Bump version number to 0.1.7